Add opt-in Box3D collider debug draw
